pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "traffic-eye"
version = "0.1.0"
description = "Edge-AI Traffic Violation Detection System"
requires-python = ">=3.11"
dependencies = [
    "pyyaml>=6.0",
    "opencv-python-headless>=4.8",
    "numpy>=1.24",
    "Pillow>=10.0",
    "jinja2>=3.1",
    "psutil>=5.9",
    "httpx>=0.25",
    "schedule>=1.2",
    "pynmea2>=1.18",
    "geopy>=2.4",
    "paddleocr>=2.7,<3.0",  # v3.x requires full PaddlePaddle (too heavy for edge)
    "google-cloud-aiplatform>=1.136",  # For Vertex AI cloud-only OCR
]

[project.optional-dependencies]
pi = [
    "tflite-runtime",
    "picamera2",
    "gps3",
]
dev = [
    "pytest>=7.0",
    "pytest-cov",
    "pytest-asyncio",
    "ruff",
    "mypy",
]

[project.scripts]
traffic-eye = "src.main:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
